Version 0.8.3.1
This is the version 0.8.3.1 of ReefAngel.  This is before any changes to the library took place.  This contains all the code in a single file with the help of additional libraries.  If you are to use this code, you will want to use the ReefAngelTutorial_v0831 Sketch file too.

This does not use any of the ReefAngel_ libraries.