ReefAngel_Wifi.cpp - Fixed error with no wavemakers in use
When setting the internal memory from the web interface, if the memory location is a Wavemaker we would restart the wavemaker timers.  If wavemakers are not in use, this code would error out due to missing / undeclared variables.  The code is now updated to only restart if a wavemaker is declared (by defining WavemakerSetup in the Features file).