Updated Feeding and Water Change Modes
Feeding and water change modes now process all the normal controller functions.  The loop() is now being processed so the controller still monitors the devices.  The web interface is now active during these modes as well.  Future improvements may be needed.  The devices that get turned on in the loop() that get toggled off for the modes may get activated by the controller.  The only devices that currently respond appropriately are Wavemakers and Always On devices (turned on in the setup() ).

ReefAngel.cpp, ReefAngel.h, ReefAngel_Globals.h