Simplified Menu Added
Added in simplified menu.  Enabled by #define SIMPLE_MENU
It's a single menu screen with the basic options in it.  Useful
for those who want the bare minimum on the controller and setup
everything elsewhere (ClientSuite, SetInternalMemory PDE, etc).
Also added in functions for Starting FeedingMode and WaterChangeMode
as well as clearing the ATO Timeout and Overheat Timeout.