Temperature Conversion
The setup menus now display the appropriate units based on the controller's unit instead of being hard coded to F.  The ranges for the temps are now changed appropriately.  There is a function that can be called to convert the internal memory values from F to C or vice versa.  It's not called by default to save space.  It should be called when we have more space to work with.  This will ensure that no matter what you set the internal memory values to, they will be converted properly based on the temperature setting on the controller.