ReefAngel_NokiaLCD - Added in fonts 8x8, 8x16 and 12x16
Added in the ability to enable the fonts for use in custom main screen
creation.  They can be enabled by the defines: FONT_8x8, FONT_8x16
and FONT_12x16 along with the NUMBERS_16x16 define
When defined, you will have a function to call to use the appropriate font
8x8 - DrawLargeText(..., 8)
8x16 - DrawLargeText(..., 16)
12x16 - DrawHugeText(...)
16x16 - DrawNumbersText(...)

The DrawLargeText has the same parameters except the last parameter
specifies the height of the font to use.  If you want the 8x8 font you
must specify 8 and if you want the 8x16 font you must specify 16.

Also, currently DrawNumbersText only works with a string, so you
cannot pass in actual numbers.  This may change but for now it must be
a string that is passed in.