Updated Font Diplaying and Functions
Merged the DrawLargeTextLine functions together for the fonts 8x8, 8x16
and 12x16.  This function is used by all three of these fonts with
the distinction being made by the last parameter which is the size.
The size is actually an enumerated value:
Font8x8
Font8x16
Font12x16

If the size is omitted, it is assumed to be Font8x8, so this could
have problems displaying the proper font if improperly used.  No way
to safeguard around this.