Added numbers only fonts
In addition to the additional fonts, since most of the fonts being displayed
are going to be used for numbers on the main screen, the corresponding
numbers to the fonts have been isolated.  This saves on memory space and
allows people to use the fonts.

The numbers available are:  8x8, 8x16, 12x16 and 16x16

DrawHugeText now takes a last parameter of a font height which is an
enum of either Font12x16 or Num12x16.