Wifi - Added in handling of Date & Time setting from wifi interface
Date & Time is now able to be set via the wifi interface.  The format is:

GET /dHHMM,MMDD,YY

HH - hours
MM - minutes
MM - month
DD - day
YY - year since 2000

All numbers MUST be formatted to 2 digits, so you must send the leading
0's in order for it to work properly.

The controller will reply <D>ERR</D> on error or <D>OK</D> on success.
There is no indication of what the error is.  Simple validation is performed
to ensure that the values passed in are within normal ranges but specifics
for all the months is not checked.  It is the sender's responsibility to
ensure that the date is validated and sent properly.