Wifi - Update how we process unknown wifi requests
Unknown request handling was added in an earlier release.  So if we do not
recognize the request, we send an Unknown Request response.  This works great
for the client suite and if people are sending commands manually.  This fails
when people use the WebBanner to update reefangel.com.  Once the controller
sends the data to reefangel.com, the server replies back with a response
and we process that as an unknown request and send back an unknown request
response.  Then ra.com gets that request and doesn't know how to handle it
and sends us another response back.  We then get in an almost endless loop
of the requests being sent back and forth and no data gets updated.

This update will mark it as an unknown request and then proceed without
sending anything back to the sender.  It should not affect anything at all
with other programs communicating via WIFI.  If anything, this should be
a little more secure in the sense that the controller is not publically or
physically responding to requests it does not handle.  So people will not
know if it was a good request or not unless they actually get a response back.