ConvertNumToString function added
The function ConvertNumToString has been added to simplify the conversion
from the numeric values to display properly in a string format.  This is
useful for the Temperatures and PH values.  It will convert the number to
a string and also insert the appropriate decimal place (provided the
proper positioning is given).