Exceed flags storing
The logging / storing of which event caused the RED status led to be
turned on.  The ATO or Overheat lights are the 2 ways the status led
is turned on.  Currently there is no way of knowing which event caused
the led to be turned on.  Enabling this feature will allow the event
to be logged and displayed on your screen.  It does not log the time,
only that either ATO or Overheat triggered the LED.

When you clear the events the LED is turned off but only the event
being cleared will be cleared.  Currently the Overheat clear and ATO
clear will turn the LED off.  With this enabled, you must clear each
event that caused the LED to be turned on.

There is no distinction between the LOW or HIGH ato switches.
There is a distinction between Standard ATO and Single ATO but both
functions cannot be used at the same time.

To know what event triggered the LED, you MUST use a custom main
screen and perform your own code checks to see the event.  Future
improvements may be incorporated elsewhere but nothing is currently
planned.

This is enabled as follows in the Features file:

"#define ENABLE_EXCEED_FLAGS"

The logging will require an additional 68 bytes of storage.