NokiaLCD - Updated Error displaying for temp sensors
DrawSingleMonitor now checks if we are adding a decimal point to the
number being displayed.  If we are, we know that it is most likely a
temperature or ph.  If the value is 0, then we will display Error
indicating that there is an error with the probe.  Otherwise, if no
decimal is being drawn, we most likely are displaying the PWM percentage
and therefore should display the 0.