Added additional Custom Menu modes
When displaying a custom menu, if you display your own screen
you need to indicate if you want to return to the menu or the
main/home screen.  The modes will handle the appropriate cleanup
and enabling of proper settings so the controller will work as
desired.
The modes / displayed menu options are:
  ALT_SCREEN_MODE - for returning to menu
  RETURN_MAIN_MODE - for returning to home screen