Wifi - added in more wifi commands
Added the following commands:
/f - starts feeding mode
/w - starts water change mode
/bp - simulates a button press

Feeding and water change modes can only be activated if you are on the
main screen.  If you are already in one of those modes, then the event
is ignored.  Also, button press works in ANY mode.  It can be used to
stop the modes or to display the menu, but you cannot mimic the
joystick.

Also removed some pingSerial calls during the drawing of the home
screen. If one of those modes was triggered during a drawing call
then the screen would have artifacts on it when the screen was changed.