Updated mode response
When we enter into the feeding or water change mode successfully,
the controller will return <MODE>OK</MODE>
If we are already in a mode or unable to enter the mode, the
controller will return <MODE>ERR</MODE> indicating that it cannot
change modes.
This applies to /mw and /mf (water and feeding modes).